requirements.txt
tinydb==4.7.0
github3.py==3.2.0
autocorrect==2.6.1
langdetect==1.0.9
python-telegram-bot==13.11
python-dotenv==0.20.0
